requirements.txt
pytz>=2015.2
django-debug-toolbar==1.3.2
celery==3.1.18
redis>=2.10.3
Django>=1.8
dj-database-url==0.3.0
elasticsearch==1.4.0
elasticstack==0.1.1
eventlet>=0.17.4
django-haystack>=2.3.1
django-localflavor==1.1
https://bitbucket.org/schinckel/django-postgres/get/ada11ed1caef.zip
psycopg2==2.6
gunicorn==19.3.0
requests>=2.7.0
whitenoise==1.0.6
django-grappelli>=2.6.5
xlrd==0.9.3
django-libsass